update quo function to accept f32 parameters and adjust unit tests accordingly

+++ scalar/src/lib.rs
@@ -10,7 +10,7 @@ pub fn pro(a: i32, b: i32) -> i32 {
     a * b
 }
 
-pub fn quo(a: i32, b: i32) -> i32 {
+pub fn quo(a: f32, b: f32) -> f32 {
     a / b
 }
 
@@ -33,8 +33,8 @@ mod tests {
         let pro_result = pro(2, 2);
         assert_eq!(pro_result, 4);
 
-        let quo_result = quo(2, 2);
-        assert_eq!(quo_result, 1);
+        let quo_result = quo(2.0, 2.0);
+        assert_eq!(quo_result, 1.0);
 
         let rem_result = rem(2, 2);
         assert_eq!(rem_result, 0);
